refactor(live-preview): split LivePreview into toolbar/stage/frames + hooks
LivePreview.tsx had grown to ~970 lines mixing UI, the mobile-iframe
lifecycle, availability probing, navigation sync and theme sync. Extract
the presentational pieces (LivePreviewToolbar, LivePreviewStage,
LivePreviewWebFrame, LivePreviewMobileFrame) and the behaviour into focused
hooks (useMobilePreviewAvailability, useMobilePreviewSession,
usePreviewNavigationSync, usePreviewPreferenceSync, usePreviewUrlMirror,
usePageVisibilityUnmount) plus a pure keywordFromPreviewPath util.

LivePreview is now the orchestration shell that owns the canonical preview
state and wires the hooks to the presentational tree. No behaviour change:
iframe URLs, one-time-code minting, the fresh-code remount lifecycle,
navigation sync, draft, theme sync and tab-hidden unmount are unchanged
(verified in-browser: device switch resizes without reload, web<->mobile
nav sync, theme sync, draft).
