+++ src/app/components/cms/live-preview/LivePreview.tsx
@@ -37,12 +37,14 @@ SPDX-License-Identifier: MPL-2.0
  *     each pane's own controls (web header selector / mobile profile), not the
  *     toolbar.
  *
- * MOBILE IFRAME LIFECYCLE (smooth in dev + prod). The mobile iframe is its own
- * HMR client in dev, so a backgrounded preview tab can starve the Expo dev
- * server. It is unloaded while the TAB IS HIDDEN; merely losing window focus
- * (DevTools, the IDE, another window) keeps it live. Returning remounts it with
- * a fresh code. Reload-resilience for the one-time code lives in the mobile
- * image (sessionStorage keyed by the code).
+ * This component is the ORCHESTRATION shell: it owns the canonical preview state
+ * (current page, draft, device/orientation, reload keys) and composes the
+ * presentational `LivePreviewToolbar` + `LivePreviewStage` with the behaviour
+ * hooks (`useMobilePreviewAvailability`, `usePageVisibilityUnmount`,
+ * `useMobilePreviewSession`, `usePreviewPreferenceSync`,
+ * `usePreviewNavigationSync`, `usePreviewUrlMirror`). The mobile-iframe lifecycle
+ * (unload while the TAB IS HIDDEN, fresh-code remount on resume/reload) lives in
+ * those hooks.
  *
  * Gated by the `admin.mobile_preview.view` permission (server-checked in the
  * route, client-checked for the editor entry point). The admin JWT never
@@ -53,40 +55,10 @@ SPDX-License-Identifier: MPL-2.0
 
 import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import { useRouter } from 'next/navigation';
-import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
-import { REACT_QUERY_CONFIG } from '../../../../config/react-query.config';
-import {
-    ActionIcon,
-    Alert,
-    Badge,
-    Box,
-    Button,
-    Divider,
-    Group,
-    Loader,
-    SegmentedControl,
-    Stack,
-    Switch,
-    Text,
-    Tooltip,
-    useMantineColorScheme,
-} from '@mantine/core';
+import { useQueryClient } from '@tanstack/react-query';
+import { Box, useMantineColorScheme } from '@mantine/core';
 import { useElementSize } from '@mantine/hooks';
-import {
-    IconAlertTriangle,
-    IconArrowLeft,
-    IconDeviceMobile,
-    IconExternalLink,
-    IconRefresh,
-} from '@tabler/icons-react';
-import {
-    PREVIEW_BRIDGE_MESSAGE,
-    arePreviewPreferencesEqual,
-    isPreviewBridgeMessage,
-    type IPreviewPreferences,
-    type TPreviewBridgeMessage,
-} from '@selfhelp/shared';
-import { AdminMobilePreviewApi } from '../../../../api/admin/mobile-preview.api';
+import { REACT_QUERY_CONFIG } from '../../../../config/react-query.config';
 import { API_CONFIG } from '../../../../config/api.config';
 import { useAppNavigation } from '../../../../hooks/useAppNavigation';
 import { usePublicLanguages } from '../../../../hooks/useLanguages';
@@ -94,20 +66,19 @@ import { useLanguageContext } from '../../contexts/LanguageContext';
 import { usePreviewMode } from '../../contexts/PreviewModeContext';
 import {
     buildMobilePreviewUrl,
-    DEFAULT_MOBILE_PREVIEW_ORIGIN,
     isAbsolutePreviewOrigin,
-    previewOriginCandidates,
-    type TPreviewDevice,
     type TPreviewModalMode,
-    type TPreviewOrientation,
 } from '../pages/mobile-preview/mobilePreviewUrl';
-import {
-    computeFrameLayout,
-    isPreviewPageActive,
-    livePreviewThemePreferences,
-} from './livePreviewLayout';
-import { LivePreviewWebPane } from './LivePreviewWebPane';
+import { computeFrameLayout } from './livePreviewLayout';
 import { useLivePreviewToolbarStore } from '../../../store/livePreview.store';
+import { LivePreviewToolbar } from './LivePreviewToolbar';
+import { LivePreviewStage } from './LivePreviewStage';
+import { useMobilePreviewAvailability } from './hooks/useMobilePreviewAvailability';
+import { usePageVisibilityUnmount } from './hooks/usePageVisibilityUnmount';
+import { useMobilePreviewSession } from './hooks/useMobilePreviewSession';
+import { usePreviewPreferenceSync } from './hooks/usePreviewPreferenceSync';
+import { usePreviewNavigationSync } from './hooks/usePreviewNavigationSync';
+import { usePreviewUrlMirror } from './hooks/usePreviewUrlMirror';
 
 /**
  * localStorage flag: the live preview defaults to DRAFT only the FIRST time it
@@ -116,54 +87,6 @@ import { useLivePreviewToolbarStore } from '../../../store/livePreview.store';
  * reset to draft on every reload.
  */
 const LIVE_PREVIEW_DRAFT_DEFAULTED_KEY = 'sh:live-preview:draft-defaulted';
-
-interface IMobilePreviewVersionInfo {
-    version?: string | null;
-    mobileRendererVersion?: string | null;
-}
-
-interface ILivePreviewAvailability {
-    available: boolean;
-    origin: string;
-    dev: boolean;
-    info: IMobilePreviewVersionInfo | null;
-}
-
-/**
- * Probe a candidate origin's `<origin>/version.json`. An `optimistic` candidate
- * (a cross-origin dev server that may not serve it) is treated as available
- * even on a non-200; a same-origin installed image requires the 200. Mirrors
- * the inspector panel's probe.
- */
-async function probeCandidate(
-    origin: string,
-    optimistic: boolean,
-): Promise<{ available: boolean; info: IMobilePreviewVersionInfo | null }> {
-    try {
-        const res = await fetch(`${origin}/version.json`, { cache: 'no-store' });
-        if (res.ok) {
-            try {
-                return { available: true, info: (await res.json()) as IMobilePreviewVersionInfo };
-            } catch {
-                return { available: true, info: null };
-            }
-        }
-        return { available: optimistic, info: null };
-    } catch {
-        return { available: optimistic, info: null };
-    }
-}
-
-function toErrorMessage(error: unknown): string {
-    if (error instanceof Error && error.message) return error.message;
-    return 'Failed to start the live preview. Please retry.';
-}
-
-/** Normalise an (origin-stripped) preview path to a CMS keyword (`null` → home). */
-function keywordFromPreviewPath(path: string): string | null {
-    const cleaned = path.split('#')[0].split('?')[0].replace(/^\/+/, '').replace(/\/+$/, '');
-    return cleaned === '' ? null : cleaned;
-}
 
 export interface ILivePreviewProps {
     /** Keyword of the page the preview is launched on (initial route). */
@@ -180,10 +103,6 @@ export function LivePreview({ keyword, modal }: ILivePreviewProps) {
     const router = useRouter();
     const explicitOrigin = process.env.NEXT_PUBLIC_MOBILE_PREVIEW_ORIGIN ?? null;
     const isDev = process.env.NODE_ENV !== 'production';
-    const candidates = useMemo(
-        () => previewOriginCandidates({ explicitOrigin, isDev }),
-        [explicitOrigin, isDev],
-    );
 
     const { languages, isLoading: languagesLoading } = usePublicLanguages();
 
@@ -225,54 +144,23 @@ export function LivePreview({ keyword, modal }: ILivePreviewProps) {
     const setOrientation = useLivePreviewToolbarStore((s) => s.setOrientation);
     const showMobile = useLivePreviewToolbarStore((s) => s.showMobile);
     const setShowMobile = useLivePreviewToolbarStore((s) => s.setShowMobile);
-    // Bumping a reload key is the ONLY thing that remounts a frame. The inline
-    // web pane remounts on `webReloadKey`; the mobile iframe on `mobileReloadKey`.
+
+    // Bumping the web reload key remounts the inline web pane (a hard refresh);
+    // the mobile iframe's own remount lifecycle lives in `useMobilePreviewSession`.
     const [webReloadKey, setWebReloadKey] = useState(0);
-    const [mobileReloadKey, setMobileReloadKey] = useState(0);
     // The shell owns the canonical preview page. `currentKeyword` follows the
     // user's navigation in EITHER pane; `mobileLoadKeyword` is the keyword the
     // mobile frame last (re)loaded with — it only changes on a reload so soft
     // sync navigation never remounts the frame.
     const [currentKeyword, setCurrentKeyword] = useState<string | null>(keyword);
     const [mobileLoadKeyword, setMobileLoadKeyword] = useState<string | null>(keyword);
-    // `pageActive` follows ONLY page visibility (the tab being hidden). The
-    // mobile iframe mounts while active, so a hidden tab releases its HMR client
-    // instead of starving the Expo dev server.
-    const [pageActive, setPageActive] = useState(true);
-    // Remount latch for the mobile frame: a reload UNMOUNTS it and the remount
-    // effect brings it back only once a FRESH code is minted — remounting onto the
-    // old (already-consumed) code, or an in-place `src` swap, can wedge the
-    // cross-origin Expo dev frame on a perpetual loading state.
-    const [mobileMounted, setMobileMounted] = useState(true);
-    const [mobileReloadPending, setMobileReloadPending] = useState(false);
 
     const mobileIframeRef = useRef<HTMLIFrameElement>(null);
     const currentKeywordRef = useRef<string | null>(keyword);
-    const pageActiveRef = useRef(true);
-    // Latest minted code (ref so the reload helpers read it without re-creating
-    // callbacks every mint) + the code we are replacing on the current reload.
-    const codeRef = useRef<string | null>(null);
-    const reloadFromCodeRef = useRef<string | null>(null);
-    // Loop guard: when the shell pushes a NAVIGATE to the mobile frame it records
-    // the keyword here; that frame's echoed NAVIGATED is then swallowed instead
-    // of bounced back (no ping-pong).
-    const expectedRef = useRef<{ mobile: string | null | undefined }>({ mobile: undefined });
-    // Loop guard for prefs sync (the theme/language analogue of expectedRef): the
-    // last prefs the shell pushed to OR received from the mobile frame. Plus a
-    // live snapshot of the web pane's prefs so the mobile READY handler can push
-    // them without the listener depending on (and re-subscribing to) them.
-    const lastSyncedPrefsRef = useRef<IPreviewPreferences | null>(null);
-    const currentPrefsRef = useRef<IPreviewPreferences>(livePreviewThemePreferences('auto'));
 
     useEffect(() => {
         currentKeywordRef.current = currentKeyword;
     }, [currentKeyword]);
-
-    useEffect(() => {
-        // Theme-only live sync. Language is bound to the iframe URL and changes via
-        // a remount (see `selectedLanguageId`), so it is never pushed over the bridge.
-        currentPrefsRef.current = livePreviewThemePreferences(colorScheme);
-    }, [colorScheme]);
 
     // Default the preview to DRAFT the FIRST time it is ever opened (parity with
     // the old default), then RESPECT the user's saved choice. The `sh_preview`
@@ -289,7 +177,7 @@ export function LivePreview({ keyword, modal }: ILivePreviewProps) {
         if (!isPreviewMode) togglePreviewMode();
     }, [isPreviewMode, togglePreviewMode]);
 
-    const { ref: bodyRef, width: bodyWidth, height: bodyHeight } = useElementSize();
+    const { ref: bodyRef, width: bodyWidth, height: bodyHeight } = useElementSize<HTMLDivElement>();
 
     // The mobile iframe is minted in — and bound to — the WEB pane's current
     // language. When the web header language selector changes it, the mint key
@@ -299,56 +187,6 @@ export function LivePreview({ keyword, modal }: ILivePreviewProps) {
     // `setLanguage`, which under the cross-frame echo looped into an invalidation
     // storm (the symptom: flooded console + an empty drawer/tab menu).
     const selectedLanguageId = currentLanguageId ?? languages[0]?.id ?? null;
-
-    // Unload the mobile iframe only while the TAB IS HIDDEN. A merely unfocused
-    // window (DevTools, the IDE) keeps it mounted, so DevTools no longer tears
-    // the preview down. On return, remount the mobile frame at the canonical page
-    // with a fresh code.
-    useEffect(() => {
-        const applyPageActive = (active: boolean) => {
-            if (pageActiveRef.current === active) return;
-            pageActiveRef.current = active;
-            setPageActive(active);
-            if (active) {
-                // Remount the mobile frame fresh on return — same safe path as a
-                // manual reload (unmount, re-mint, remount only once code is ready).
-                reloadFromCodeRef.current = codeRef.current;
-                setMobileLoadKeyword(currentKeywordRef.current);
-                setMobileMounted(false);
-                setMobileReloadPending(true);
-                setMobileReloadKey((k) => k + 1);
-            }
-        };
-
-        const onVisibility = () => {
-            applyPageActive(isPreviewPageActive({ visibilityState: document.visibilityState }));
-        };
-
-        onVisibility();
-        document.addEventListener('visibilitychange', onVisibility);
-        return () => {
-            document.removeEventListener('visibilitychange', onVisibility);
-        };
-    }, []);
-
-    // Mirror the synced page into the shell's own address bar
-    // (`/admin/preview/<keyword>`), WITHOUT a Next navigation (history API only,
-    // so the LivePreview component never remounts). Makes a manual reload restart
-    // at the page you navigated to, and the URL shareable — in BOTH directions
-    // (web in-pane nav AND mobile-reported nav both update `currentKeyword`).
-    useEffect(() => {
-        if (typeof window === 'undefined') return;
-        const kw = currentKeyword?.trim() ? currentKeyword.trim().replace(/^\/+/, '') : '';
-        const path = kw === ''
-            ? '/admin/preview'
-            : `/admin/preview/${kw.split('/').map((seg) => encodeURIComponent(seg)).join('/')}`;
-        const next = `${path}${window.location.search}${window.location.hash}`;
-        const current = `${window.location.pathname}${window.location.search}${window.location.hash}`;
-        if (next !== current) {
-            window.history.replaceState(window.history.state, '', next);
-        }
-    }, [currentKeyword]);
-
     const locale = useMemo(
         () =>
             ctxLanguages.find((lang) => lang.id === selectedLanguageId)?.locale ??
@@ -358,46 +196,12 @@ export function LivePreview({ keyword, modal }: ILivePreviewProps) {
     );
 
     // --- availability probe (React Query owns the state) -----------------------
-    const availabilityQuery = useQuery<ILivePreviewAvailability>({
-        queryKey: ['live-preview-version', candidates.map((c) => `${c.mode}:${c.origin}`).join('|')],
-        queryFn: async () => {
-            for (const candidate of candidates) {
-                if (candidate.optimistic) {
-                    return {
-                        available: true,
-                        origin: candidate.origin,
-                        dev: candidate.mode === 'dev',
-                        info: null,
-                    };
-                }
-                if (isDev && candidate.mode === 'installed') {
-                    continue;
-                }
-                const result = await probeCandidate(candidate.origin, candidate.optimistic);
-                if (result.available) {
-                    return {
-                        available: true,
-                        origin: candidate.origin,
-                        dev: candidate.mode === 'dev',
-                        info: result.info,
-                    };
-                }
-            }
-            const fallback = candidates[0]?.origin ?? DEFAULT_MOBILE_PREVIEW_ORIGIN;
-            return { available: false, origin: fallback, dev: false, info: null };
-        },
-        staleTime: 30_000,
-        retry: false,
-    });
+    const { availability, previewOrigin, devOrigin, versionInfo, refetch } =
+        useMobilePreviewAvailability({ explicitOrigin, isDev });
 
-    const availability: 'checking' | 'available' | 'unavailable' = availabilityQuery.isLoading
-        ? 'checking'
-        : availabilityQuery.data?.available
-          ? 'available'
-          : 'unavailable';
-    const previewOrigin = availabilityQuery.data?.origin ?? candidates[0]?.origin ?? DEFAULT_MOBILE_PREVIEW_ORIGIN;
-    const devOrigin = availabilityQuery.data?.dev ?? false;
-    const versionInfo = availabilityQuery.data?.info ?? null;
+    // `pageActive` follows ONLY page visibility (the tab being hidden); a hidden
+    // tab unmounts the mobile iframe so it stops starving the Expo dev server.
+    const { pageActive } = usePageVisibilityUnmount();
 
     // The mobile iframe mounts (and the code is minted) only while the preview is
     // available and the page is active. The WEB pane is inline → always rendered.
@@ -416,64 +220,40 @@ export function LivePreview({ keyword, modal }: ILivePreviewProps) {
         }
     }, [previewOrigin]);
 
-    // --- mint (keyword-less → free navigation) ---------------------------------
-    const mintMutation = useMutation({
-        mutationFn: () =>
-            AdminMobilePreviewApi.createSession({
-                language_id: selectedLanguageId ?? undefined,
-                draft,
-            }),
-        // Auto-recover from a transient failure (cold dev route-compile can exceed
-        // the client timeout on the first mint; the backend may briefly 503
-        // mid-restart). Minted codes are single-use + cheap; the error UI is the
-        // fallback once these are exhausted.
-        retry: 2,
-        retryDelay: (attempt) => Math.min(1000 * 2 ** attempt, 4000),
-    });
-    const { mutate: mintCode } = mintMutation;
-    const code = mintMutation.data?.code ?? null;
-    const mintError = mintMutation.isError ? toErrorMessage(mintMutation.error) : null;
+    // --- mobile session: mint (keyword-less → free navigation) + remount --------
+    const { code, mintError, mintPending, mobileMounted, reloadMobileFresh } =
+        useMobilePreviewSession({
+            previewActive,
+            pageActive,
+            languagesLoading,
+            languagesCount: languages.length,
+            selectedLanguageId,
+            draft,
+            previewOrigin,
+            currentKeywordRef,
+            setMobileLoadKeyword,
+        });
 
-    useEffect(() => {
-        codeRef.current = code;
-    }, [code]);
-
-    // Remount the mobile iframe ONLY once a FRESH code has been minted after a
-    // reload/refresh/resume. Bringing it back on the old, already-consumed code —
-    // or swapping the `src` in place — wedges the cross-origin Expo dev frame on a
-    // perpetual loading spinner (the symptom: it only "unsticks" after the tab is
-    // hidden + shown). Waiting for the new code means a single clean mount.
-    useEffect(() => {
-        if (!mobileReloadPending) return;
-        if (!code || code === reloadFromCodeRef.current) return;
-        setMobileReloadPending(false);
-        setMobileMounted(true);
-    }, [mobileReloadPending, code]);
-
-    // Mint exactly ONCE per distinct intent (dedup against StrictMode double
-    // invoke + independent availability/language transitions). Re-mints only when
-    // something that affects the token changes: language, draft, or a reload.
-    const lastMintKeyRef = useRef<string | null>(null);
-    useEffect(() => {
-        if (!previewActive) return undefined;
-        if (languagesLoading) return undefined;
-        if (languages.length > 0 && selectedLanguageId === null) return undefined;
-
-        const key = `${previewOrigin}|${selectedLanguageId ?? ''}|${draft ? 1 : 0}|${mobileReloadKey}`;
-        if (lastMintKeyRef.current === key) return undefined;
-        lastMintKeyRef.current = key;
-        mintCode();
-        return undefined;
-    }, [
+    // --- cross-pane theme sync + synchronized navigation -----------------------
+    const { sendPreferencesMobile, currentPrefsRef } = usePreviewPreferenceSync({
         previewActive,
-        languagesLoading,
-        languages.length,
-        selectedLanguageId,
-        draft,
-        mobileReloadKey,
-        previewOrigin,
-        mintCode,
-    ]);
+        colorScheme,
+        setColorScheme,
+        mobileMessageOrigin,
+        mobileIframeRef,
+    });
+    const { handleWebNavigate } = usePreviewNavigationSync({
+        previewActive,
+        mobileMessageOrigin,
+        mobileIframeRef,
+        currentKeywordRef,
+        setCurrentKeyword,
+        currentPrefsRef,
+        sendPreferencesMobile,
+    });
+
+    // Mirror the canonical page into the shell's own address bar (history only).
+    usePreviewUrlMirror(currentKeyword);
 
     // Device / orientation are NOT in the URL: the iframe is sized to the device
     // and the app renders responsively, so rotating/resizing never reloads it.
@@ -523,16 +303,6 @@ export function LivePreview({ keyword, modal }: ILivePreviewProps) {
         [device, orientation, bodyWidth, bodyHeight, frameChromeHeight],
     );
 
-    // Reload the mobile frame the SAFE way: unmount it, mint a FRESH code, and let
-    // the remount effect bring it back only once that new code is ready (avoids the
-    // perpetual-loading wedge of remounting onto a consumed code / in-place swap).
-    const reloadMobileFresh = useCallback(() => {
-        reloadFromCodeRef.current = codeRef.current;
-        setMobileLoadKeyword(currentKeywordRef.current);
-        setMobileMounted(false);
-        setMobileReloadPending(true);
-        setMobileReloadKey((k) => k + 1);
-    }, []);
     const handleReloadMobile = reloadMobileFresh;
     const handleRefresh = useCallback(() => {
         // The web pane renders cached page content, so a remount alone replays the
@@ -552,126 +322,6 @@ export function LivePreview({ keyword, modal }: ILivePreviewProps) {
         },
         [reloadMobileFresh, setShowMobile],
     );
-
-    // Push a soft "navigate to keyword" to the mobile frame (no reload) and record
-    // it as expected so its echoed NAVIGATED isn't bounced back (loop guard).
-    const sendNavigateMobile = useCallback(
-        (kw: string | null) => {
-            const win = mobileIframeRef.current?.contentWindow;
-            if (!win || !mobileMessageOrigin) return;
-            expectedRef.current.mobile = kw;
-            const message: TPreviewBridgeMessage = {
-                type: PREVIEW_BRIDGE_MESSAGE.NAVIGATE,
-                keyword: kw,
-            };
-            win.postMessage(message, mobileMessageOrigin);
-        },
-        [mobileMessageOrigin],
-    );
-
-    // Push the shared THEME down to the mobile frame (no reload) and record it as
-    // the last synced value so the frame's echoed PREFERENCES_CHANGED isn't bounced
-    // back (loop guard, mirror of sendNavigateMobile). Language is NOT pushed here —
-    // it is bound to the iframe URL and changes via a remount (see selectedLanguageId).
-    const sendPreferencesMobile = useCallback(
-        (prefs: IPreviewPreferences) => {
-            const win = mobileIframeRef.current?.contentWindow;
-            if (!win || !mobileMessageOrigin) return;
-            const themePrefs = livePreviewThemePreferences(prefs.colorScheme);
-            lastSyncedPrefsRef.current = themePrefs;
-            const message: TPreviewBridgeMessage = {
-                type: PREVIEW_BRIDGE_MESSAGE.SET_PREFERENCES,
-                preferences: themePrefs,
-            };
-            win.postMessage(message, mobileMessageOrigin);
-        },
-        [mobileMessageOrigin],
-    );
-
-    // Web pane → mobile: when the web pane's THEME changes (its header ThemeToggle),
-    // relay it to the mobile frame. The guard skips the echo of a value the mobile
-    // just reported up. (Language is handled by the iframe remount, not here.)
-    useEffect(() => {
-        if (!previewActive) return;
-        const prefs = livePreviewThemePreferences(colorScheme);
-        if (arePreviewPreferencesEqual(lastSyncedPrefsRef.current, prefs)) return;
-        sendPreferencesMobile(prefs);
-    }, [colorScheme, previewActive, sendPreferencesMobile]);
-
-    // In-pane web navigation (intercepted links/buttons) → make it the canonical
-    // page and drive the mobile frame to match.
-    const handleWebNavigate = useCallback(
-        (path: string) => {
-            const kw = keywordFromPreviewPath(path);
-            if (kw === currentKeywordRef.current) return;
-            currentKeywordRef.current = kw;
-            setCurrentKeyword(kw);
-            sendNavigateMobile(kw);
-        },
-        [sendNavigateMobile],
-    );
-
-    // Synchronized navigation FROM mobile: when the mobile frame reports it
-    // navigated, make it the canonical page (which re-renders the inline web pane
-    // and mirrors the URL). The loop guard stops our own pushed command's echo
-    // from bouncing.
-    useEffect(() => {
-        if (!previewActive) return undefined;
-        const onMessage = (event: MessageEvent) => {
-            if (event.origin !== mobileMessageOrigin) return;
-            if (!isPreviewBridgeMessage(event.data)) return;
-            const data = event.data;
-            // The mobile frame announces READY after a (re)load → push the
-            // canonical page AND the current theme so it syncs to wherever the web
-            // pane already is (web pane wins the initial sync). Language is already
-            // baked into the frame's URL at mint, so it isn't pushed here.
-            if (data.type === PREVIEW_BRIDGE_MESSAGE.READY) {
-                sendNavigateMobile(currentKeywordRef.current);
-                sendPreferencesMobile(currentPrefsRef.current);
-                return;
-            }
-            if (data.type !== PREVIEW_BRIDGE_MESSAGE.NAVIGATED) return;
-            if (data.source !== 'mobile') return;
-
-            // Echo of a command we pushed → consume + stop.
-            if (expectedRef.current.mobile === data.keyword) {
-                expectedRef.current.mobile = undefined;
-                return;
-            }
-            // Already on this page (covers the initial boot echo) → no update.
-            if (data.keyword === currentKeywordRef.current) return;
-
-            currentKeywordRef.current = data.keyword;
-            setCurrentKeyword(data.keyword);
-        };
-        window.addEventListener('message', onMessage);
-        return () => {
-            window.removeEventListener('message', onMessage);
-        };
-    }, [previewActive, mobileMessageOrigin, sendNavigateMobile, sendPreferencesMobile]);
-
-    // Mobile → web pane: a THEME change reported BY the mobile frame is applied to
-    // the SAME Mantine scheme the web pane uses, so the web pane follows. Separate
-    // listener so it can depend on the live scheme/setter without re-subscribing
-    // the navigation listener. The guard skips the echo of a value the shell just
-    // pushed down. Language is NOT adopted from the mobile frame — it is driven
-    // one-way (web → mobile) via the iframe mint/remount, so there is no language
-    // echo for the two panes to ping-pong on.
-    useEffect(() => {
-        if (!previewActive) return undefined;
-        const onPrefsMessage = (event: MessageEvent) => {
-            if (event.origin !== mobileMessageOrigin) return;
-            if (!isPreviewBridgeMessage(event.data)) return;
-            if (event.data.type !== PREVIEW_BRIDGE_MESSAGE.PREFERENCES_CHANGED) return;
-            if (event.data.source !== 'mobile') return;
-            const prefs = event.data.preferences;
-            if (arePreviewPreferencesEqual(lastSyncedPrefsRef.current, prefs)) return;
-            lastSyncedPrefsRef.current = livePreviewThemePreferences(prefs.colorScheme);
-            if (prefs.colorScheme !== colorScheme) setColorScheme(prefs.colorScheme);
-        };
-        window.addEventListener('message', onPrefsMessage);
-        return () => window.removeEventListener('message', onPrefsMessage);
-    }, [previewActive, mobileMessageOrigin, colorScheme, setColorScheme]);
 
     // Back-to-editor targets the page you are CURRENTLY viewing — synced
     // navigation can move the preview off the launch keyword.
@@ -700,270 +350,46 @@ export function LivePreview({ keyword, modal }: ILivePreviewProps) {
 
     return (
         <Box style={{ display: 'flex', flexDirection: 'column', height: '100vh', overflow: 'hidden' }}>
-            {/* Top control bar. LEFT: identity + the shared controls (Mobile,
-                Draft, refresh both, open in new tab). RIGHT: the mobile device
-                controls (device, orientation, mobile-only reload) — shown only
-                while the mobile pane is on and available. */}
-            <Stack
-                gap={0}
-                style={{
-                    borderBottom: '1px solid var(--mantine-color-default-border)',
-                    background: 'var(--mantine-color-body)',
+            <LivePreviewToolbar
+                onBack={() => router.push(editorHref)}
+                currentKeyword={currentKeyword}
+                versionInfo={versionInfo}
+                devOrigin={devOrigin}
+                showMobile={showMobile}
+                onToggleMobile={handleToggleMobile}
+                draft={draft}
+                onToggleDraft={() => togglePreviewMode()}
+                onRefresh={handleRefresh}
+                refreshing={mintPending}
+                webOpenUrl={webOpenUrl}
+                availability={availability}
+                device={device}
+                onDeviceChange={setDevice}
+                orientation={orientation}
+                onOrientationChange={setOrientation}
+                onReloadMobile={handleReloadMobile}
+            />
+            <LivePreviewStage
+                bodyRef={bodyRef}
+                web={{ webReloadKey, keyword: currentKeyword, onNavigate: handleWebNavigate }}
+                showMobile={showMobile}
+                mobile={{
+                    availability,
+                    previewOrigin,
+                    isDev,
+                    onRetryAvailability: refetch,
+                    frame,
+                    bezelPadding,
+                    device,
+                    mobileUrl,
+                    mobileMounted,
+                    previewActive,
+                    mobileIframeRef,
+                    mintError,
+                    onReloadMobile: handleReloadMobile,
+                    mintPending,
                 }}
-            >
-                <Group justify="space-between" align="center" px="md" py="xs" wrap="nowrap">
-                    <Group gap="sm" align="center" wrap="nowrap">
-                        <Tooltip label="Back to editor">
-                            <ActionIcon
-                                variant="subtle"
-                                onClick={() => router.push(editorHref)}
-                                aria-label="Back to page editor"
-                            >
-                                <IconArrowLeft size="1.1rem" />
-                            </ActionIcon>
-                        </Tooltip>
-                        <IconDeviceMobile size="1.1rem" />
-                        <Text fw={600} size="sm">
-                            Live preview
-                        </Text>
-                        <Badge size="sm" variant="light" color="blue">
-                            {currentKeyword || 'home'}
-                        </Badge>
-                        {versionInfo?.version && (
-                            <Badge size="sm" variant="light" color="grape">
-                                preview v{versionInfo.version}
-                            </Badge>
-                        )}
-                        {devOrigin && (
-                            <Badge size="sm" variant="light" color="teal">
-                                live-reload dev
-                            </Badge>
-                        )}
-                        <Divider orientation="vertical" />
-                        <Switch
-                            size="xs"
-                            checked={showMobile}
-                            onChange={(e) => handleToggleMobile(e.currentTarget.checked)}
-                            label="Mobile"
-                            aria-label="Show the mobile pane"
-                        />
-                        <Switch
-                            size="xs"
-                            checked={draft}
-                            onChange={() => togglePreviewMode()}
-                            label="Draft"
-                            aria-label="Preview unpublished draft content (web and mobile)"
-                        />
-                        <Tooltip label="Refresh both previews">
-                            <ActionIcon
-                                size="md"
-                                variant="light"
-                                onClick={handleRefresh}
-                                loading={mintMutation.isPending}
-                                aria-label="Refresh all previews"
-                            >
-                                <IconRefresh size="1rem" />
-                            </ActionIcon>
-                        </Tooltip>
-                        <Tooltip label="Open the current page in a new tab">
-                            <ActionIcon
-                                size="md"
-                                variant="subtle"
-                                component="a"
-                                href={webOpenUrl}
-                                target="_blank"
-                                rel="noopener noreferrer"
-                                aria-label="Open the current page in a new tab"
-                            >
-                                <IconExternalLink size="1rem" />
-                            </ActionIcon>
-                        </Tooltip>
-                    </Group>
-
-                    {showMobile && availability === 'available' && (
-                        <Group gap="xs" align="center" wrap="nowrap" justify="flex-end">
-                            <SegmentedControl
-                                size="xs"
-                                value={device}
-                                onChange={(v) => setDevice(v as TPreviewDevice)}
-                                data={[
-                                    { label: 'Phone', value: 'phone' },
-                                    { label: 'Tablet', value: 'tablet' },
-                                ]}
-                                aria-label="Preview device"
-                            />
-                            <SegmentedControl
-                                size="xs"
-                                value={orientation}
-                                onChange={(v) => setOrientation(v as TPreviewOrientation)}
-                                data={[
-                                    { label: 'Portrait', value: 'portrait' },
-                                    { label: 'Landscape', value: 'landscape' },
-                                ]}
-                                aria-label="Preview orientation"
-                            />
-                            <Tooltip label="Reload the mobile preview only">
-                                <ActionIcon
-                                    size="md"
-                                    variant="subtle"
-                                    onClick={handleReloadMobile}
-                                    loading={mintMutation.isPending}
-                                    aria-label="Reload mobile preview"
-                                >
-                                    <IconRefresh size="1rem" />
-                                </ActionIcon>
-                            </Tooltip>
-                        </Group>
-                    )}
-                </Group>
-            </Stack>
-
-            {/* Body — inline web pane (always) + optional mobile device frame. */}
-            <Box
-                ref={bodyRef}
-                style={{
-                    flex: 1,
-                    display: 'flex',
-                    gap: 16,
-                    padding: 16,
-                    overflow: 'hidden',
-                    background: 'var(--mantine-color-default-hover)',
-                }}
-            >
-                <Box style={{ flex: 1, minWidth: 0 }}>
-                    <LivePreviewWebPane
-                        key={`web-${webReloadKey}`}
-                        keyword={currentKeyword}
-                        onNavigate={handleWebNavigate}
-                    />
-                </Box>
-
-                {showMobile && (
-                    <Box
-                        style={{
-                            flex: '0 0 auto',
-                            minWidth: 0,
-                            display: 'flex',
-                            flexDirection: 'column',
-                            alignItems: 'center',
-                            justifyContent: 'center',
-                        }}
-                    >
-                        {availability === 'checking' ? (
-                            <Group gap="xs" align="center" h="100%" px="md">
-                                <Loader size="sm" />
-                                <Text size="sm" c="dimmed">
-                                    Checking mobile preview…
-                                </Text>
-                            </Group>
-                        ) : availability === 'unavailable' ? (
-                            <Box maw={340}>
-                                <Alert
-                                    icon={<IconAlertTriangle size="1rem" />}
-                                    color="yellow"
-                                    title="Mobile preview unavailable"
-                                >
-                                    <Stack gap="xs">
-                                        <Text size="sm">
-                                            No mobile preview is running at <code>{previewOrigin}</code>.
-                                            Enable the <code>selfhelp-mobile-preview</code> service
-                                            {isDev ? (
-                                                <>
-                                                    , or start the Expo dev server (
-                                                    <code>npx expo start --web</code>)
-                                                </>
-                                            ) : null}
-                                            .
-                                        </Text>
-                                        <Group>
-                                            <ActionIcon
-                                                variant="light"
-                                                onClick={() => void availabilityQuery.refetch()}
-                                                aria-label="Retry mobile preview"
-                                            >
-                                                <IconRefresh size="1rem" />
-                                            </ActionIcon>
-                                        </Group>
-                                    </Stack>
-                                </Alert>
-                            </Box>
-                        ) : (
-                            // A realistic device bezel around the scaled screen: a
-                            // dark rounded frame + drop shadow, mirroring the mobile
-                            // app's own PhoneFrame so the preview reads like a phone
-                            // rather than a flat web card. The screen height is sized
-                            // via `frameChromeHeight` above so the bezel never clips
-                            // the bottom of the device.
-                            <Box
-                                style={{
-                                    padding: bezelPadding,
-                                    borderRadius: device === 'phone' ? 44 : 34,
-                                    background: 'linear-gradient(160deg, #2b3742 0%, #161b21 100%)',
-                                    boxShadow: '0 18px 48px rgba(0, 0, 0, 0.45)',
-                                }}
-                            >
-                                <Box
-                                    style={{
-                                        width: frame.displayWidth,
-                                        height: frame.displayHeight,
-                                        overflow: 'hidden',
-                                        borderRadius: device === 'phone' ? 34 : 24,
-                                        background: 'var(--mantine-color-body)',
-                                    }}
-                                >
-                                    {mobileUrl && mobileMounted && previewActive ? (
-                                        <iframe
-                                            key={mobileUrl}
-                                            ref={mobileIframeRef}
-                                            title="Mobile live preview"
-                                            src={mobileUrl}
-                                            sandbox="allow-scripts allow-same-origin allow-forms allow-popups"
-                                            style={{
-                                                width: frame.width,
-                                                height: frame.height,
-                                                border: 0,
-                                                transform: `scale(${frame.scale})`,
-                                                transformOrigin: 'top left',
-                                            }}
-                                        />
-                                    ) : mintError ? (
-                                        <Stack align="center" justify="center" h="100%" gap="xs" p="md">
-                                            <IconAlertTriangle
-                                                size="1.4rem"
-                                                color="var(--mantine-color-red-6)"
-                                            />
-                                            <Text size="sm" fw={600} ta="center">
-                                                Could not start the mobile preview
-                                            </Text>
-                                            <Text size="xs" c="dimmed" ta="center">
-                                                {mintError}
-                                            </Text>
-                                            <Button
-                                                size="xs"
-                                                variant="light"
-                                                leftSection={<IconRefresh size="0.9rem" />}
-                                                onClick={handleReloadMobile}
-                                                loading={mintMutation.isPending}
-                                            >
-                                                Retry
-                                            </Button>
-                                        </Stack>
-                                    ) : (
-                                        <Stack align="center" justify="center" h="100%" gap="xs" p="md">
-                                            <Loader size="sm" />
-                                            <Text size="xs" c="dimmed" ta="center">
-                                                {isDev
-                                                    ? 'Starting the mobile preview… the first load compiles the Expo dev bundle and can take a moment.'
-                                                    : 'Starting the mobile preview…'}
-                                            </Text>
-                                        </Stack>
-                                    )}
-                                </Box>
-                            </Box>
-                        )}
-                    </Box>
-                )}
-            </Box>
+            />
         </Box>
     );
 }
